+++ src/test/txlookup_tests.cpp
@@ -19,7 +19,7 @@ BOOST_AUTO_TEST_CASE(non_ctor_lookup)
 
     for (size_t i = 0; i < 100; i += 10)
     {
-        BOOST_CHECK_EQUAL(i, FindTxPosition(block, block.vtx[i]->GetHash(), false));
+        BOOST_CHECK_EQUAL(i, static_cast<size_t>(FindTxPosition(block, block.vtx[i]->GetHash(), false)));
     }
 }
 
@@ -35,7 +35,7 @@ BOOST_AUTO_TEST_CASE(ctor_lookup)
 
     for (size_t i = 0; i < 100; i += 10)
     {
-        BOOST_CHECK_EQUAL(i, FindTxPosition(block, block.vtx[i]->GetHash(), true));
+        BOOST_CHECK_EQUAL(i, static_cast<size_t>(FindTxPosition(block, block.vtx[i]->GetHash(), true)));
     }
 }
